Rename build to build_image
