+++ airbyte-ci/connectors/pipelines/pipelines/airbyte_ci/connectors/build_image/__init__.py
