+++ airbyte-ci/connectors/pipelines/tests/test_commands/test_groups/test_connectors.py
@@ -5,7 +5,7 @@
 from typing import Callable
 
 import click
-import pipelines.airbyte_ci.connectors.build.commands
+import pipelines.airbyte_ci.connectors.build_image.commands
 import pipelines.airbyte_ci.connectors.commands
 import pipelines.airbyte_ci.connectors.publish.commands
 import pipelines.airbyte_ci.connectors.test.commands
@@ -14,7 +14,7 @@ from click.testing import CliRunner
 from connector_ops.utils import METADATA_FILE_NAME, ConnectorLanguage
 from pipelines.helpers.connectors.modifed import ConnectorWithModifiedFiles
 from pipelines.airbyte_ci.connectors import commands as connectors_commands
-from pipelines.airbyte_ci.connectors.build import commands as connectors_build_command
+from pipelines.airbyte_ci.connectors.build_image import commands as connectors_build_command
 from pipelines.airbyte_ci.connectors.publish import commands as connectors_publish_command
 from pipelines.airbyte_ci.connectors.test import commands as connectors_test_command
 from tests.utils import pick_a_random_connector
